refactor: 연관성 있는 모델끼리 패키징
