+++ src/main/kotlin/lotto/model/Ticket.kt
@@ -1,5 +1,8 @@
 package lotto.model
 
+import lotto.model.lottery.Lottery
+import lotto.model.puchaseinformation.PurchaseInformation
+
 data class Ticket(
     val userLotteries: List<Lottery>,
     val purchaseInformation: PurchaseInformation,
