+++ src/main/kotlin/lotto/model/lottery/generator/LotteriesGenerationStrategy.kt
@@ -1,6 +1,6 @@
-package lotto.model.generator
+package lotto.model.lottery.generator
 
-import lotto.model.Lottery
+import lotto.model.lottery.Lottery
 
 interface LotteriesGenerationStrategy {
     fun issueLotteries(): List<Lottery>
