+++ src/main/kotlin/lotto/model/lottery/generator/LotteryNumberGenerator.kt
@@ -1,6 +1,6 @@
-package lotto.model.generator
+package lotto.model.lottery.generator
 
-import lotto.model.LotteryNumber
+import lotto.model.lottery.LotteryNumber
 
 interface LotteryNumberGenerator {
     fun generateNumbers(): Set<LotteryNumber>
